fix(NFTScan): mf-2139 update according to API updates
