+++ packages/mask/src/plugins/Trader/trending/usePriceStats.ts
@@ -3,9 +3,9 @@ import type { Currency } from '../types/index.js'
 import type { DataProvider } from '@masknet/public-api'
 import { isUndefined } from 'lodash-unified'
 import { PluginTraderRPC } from '../messages.js'
-import { Days } from '../SNSAdaptor/trending/PriceChartDaysControl.js'
 import { useChainId } from '@masknet/plugin-infra/web3'
 import { NetworkPluginID } from '@masknet/web3-shared-base'
+import { Days } from '@masknet/web3-providers'
 
 interface Options {
     coinId?: string
