+++ packages/web3-providers/src/common.ts
@@ -1,0 +1,8 @@
+export enum Days {
+    MAX = 0,
+    ONE_DAY = 1,
+    ONE_WEEK = 7,
+    ONE_MONTH = 30,
+    THREE_MONTHS = 90,
+    ONE_YEAR = 365,
+}
