+++ packages/web3-providers/src/index.ts
@@ -93,3 +93,5 @@ export const NFTScanNonFungibleTokenSolana = new NFTScanNonFungibleTokenAPI_Sola
 export const ChainbaseHistory = new ChainbaseHistoryAPI()
 export const ChainbaseFungibleToken = new ChainbaseFungibleTokenAPI()
 export const ChainbaseNonFungibleToken = new ChainbaseNonFungibleTokenAPI()
+
+export * from './common'
